refactor: input의 헨들러 제거
